Add a deprecated options array
With the config system, maintaining backwards-compatability
can be quite difficult. Especially with things like the Web
control config-write ability.
Furthermore, most config options are either totally removed
or succeeded by a different set of options. This does not require
any backwards-compatability.

The `dep_config_params` array is checked when a config option
is unknown to the current system (`config_params`).
If the option is found, a warning is logged for the user to see.
We tell them the last version it was used and why it changed.