Prevent preview duplicate if mis-configured
When running multiple cameras or a single camera with the
config `camera` lines, motion creates a thread for each
camera (even if just the one).
A mis-configuration can easily happen where the `stream_port`
option remains set in `motion.conf`. This causes the Web control
preview section to show a duplicate of one stream.
Prevent this by knowing that if we are multi-threaded,
the main thread has no active camera and don't loop over it.