Add backwards compatibility to deprecated options.
Allow setting a 'conf_value' and 'copy_func' to the deprecated option.
This way some backwards compatibility can be obtained by simply
setting the new option to the wanted value. The deprecated option
will not exist within the config structure, so it won't be able
to accidently print out in a config write.